A DatabaseAdapter compiler
A simple compiler to generate a PostgreSQL datbase schema from an
annotated Java model.

Tables and composite types are compiled from Java model objects
which are annotated.

Annotated methods on a DatabaseAdapter are compiled to PostgreSQL
functions.

A concrete DatabaseAdapter implementation is compiled at runtime
with bindings to the PostgreSQL functions.

This is the initial commit of the project, with some basic features.
Planned features include PL/Proxy support.